# feat: lazy load CLI command actions for improved startup performance