… CHECKNODEVERSION_OPTIMIZATION.md
… benchmark-results-feat-lazy-load-cli-commands-987353a43.json
… benchmark-results-main-9a44b0bfe.json
… src/commands/cache/cacheAction.ts
… src/commands/config.ts
… src/commands/config/configAction.ts
… src/commands/debug.ts
… src/commands/debug/debugAction.ts
… src/commands/delete.ts
… src/commands/eval.ts
… src/commands/eval/evalAction.ts
… src/commands/export.ts
… src/commands/export/exportAction.ts
… src/commands/generate/actions/assertionsAction.ts
… src/commands/generate/actions/datasetAction.ts
… src/commands/generate/assertions.ts
… src/commands/generate/dataset.ts
… src/commands/import.ts
… src/commands/import/importAction.ts
… src/commands/init.ts
… src/commands/list.ts
… src/commands/list/listAction.ts
… src/commands/modelScan.ts
… src/commands/modelScan/modelScanAction.ts
… src/commands/share.ts
… src/commands/share/shareAction.ts
… src/commands/show.ts
… src/commands/show/showAction.ts
… src/commands/validate.ts
… src/commands/validate/validateAction.ts