## Summary

This PR implements lazy loading for CLI command actions as outlined in TODO.md, resulting in a **15.7% overall improvement** in CLI startup performance.

### Key Changes:
- Separated command registration (lightweight) from action execution (heavyweight)
- Applied dynamic imports for all command actions
- Optimized checkNodeVersion to remove heavy imports
- Kept main.ts completely unchanged as required

## Performance Results

### Overall Performance
- **Main branch average:** 971.84ms
- **Feature branch average:** 819.00ms
- **Improvement:** 152.84ms (15.7% faster)

### Top 5 Most Improved Commands

| Command | Main (ms) | Feature (ms) | Improvement (ms) | % Faster |
|---------|-----------|--------------|------------------|----------|
| validate | 988.45 | 820.17 | 168.28 | 17.0% |
| init | 998.59 | 831.59 | 167.00 | 16.7% |
| export | 991.45 | 826.09 | 165.36 | 16.7% |
| show | 990.22 | 826.54 | 163.68 | 16.5% |
| share | 985.84 | 823.22 | 162.62 | 16.5% |

### All Commands Performance Comparison

| Command | Main (ms) | Feature (ms) | Improvement (ms) |
|---------|-----------|--------------|------------------|
| help | 950.48 | 803.65 | 146.83 |
| eval | 965.16 | 812.03 | 153.13 |
| eval help | 962.44 | 809.09 | 153.35 |
| init | 998.59 | 831.59 | 167.00 |
| view | 961.37 | 807.18 | 154.19 |
| list | 965.70 | 809.64 | 156.06 |
| show | 990.22 | 826.54 | 163.68 |
| auth login | 965.89 | 806.39 | 159.50 |
| auth logout | 965.12 | 808.45 | 156.67 |
| auth whoami | 960.51 | 808.14 | 152.37 |
| cache clear | 973.60 | 822.12 | 151.48 |
| config show | 967.38 | 812.36 | 155.02 |
| delete | 963.28 | 810.41 | 152.87 |
| export | 991.45 | 826.09 | 165.36 |
| import | 959.79 | 809.01 | 150.78 |
| share | 985.84 | 823.22 | 162.62 |
| validate | 988.45 | 820.17 | 168.28 |
| debug | 969.55 | 819.36 | 150.19 |
| model-scan | 975.96 | 823.60 | 152.36 |
| generate dataset | 989.25 | 833.66 | 155.59 |
| generate assertions | 980.55 | 827.68 | 152.87 |

## Commands Refactored

- ✅ eval
- ✅ init  
- ✅ view
- ✅ generate (dataset, assertions)
- ✅ share
- ✅ show
- ✅ list
- ✅ cache
- ✅ config
- ✅ auth
- ✅ delete
- ✅ export
- ✅ import
- ✅ validate
- ✅ debug
- ✅ modelScan

## Testing

All CI checks pass:
- ✅ Build
- ✅ Lint
- ✅ Format
- ✅ Tests
- ✅ Circular dependencies check
- ✅ Python tests
- ✅ Integration tests

## Breaking Changes

None - all changes are internal optimizations that maintain the same external API.

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>