@mldangelo 2025-07-30
## Fixed All Issues Identified by Cursor Bot ✅

I've addressed all the command signature issues and merged the latest changes from main:

### Fixes Applied:

1. **✅ Validate command** - Now accepts optional filePath, restored backward compatibility
2. **✅ Async main** - Properly awaited to prevent premature termination
3. **✅ Show command** - ID is now optional (defaults to 'latest'), restored subcommands
4. **✅ List command** - Restored prompts and datasets subcommands
5. **✅ Assertions command** - Fixed options and passes defaultConfig correctly
6. **✅ Share command** - Handles optional showAuth parameter
7. **✅ Dataset command** - Passes defaultConfig and defaultConfigPath
8. **✅ Merged latest main** - Successfully merged all changes from main branch

### Test Results:

- ✅ Linting passes
- ✅ Build succeeds
- ✅ CLI commands work correctly
- ✅ Backward compatibility maintained

The lazy loading optimizations remain intact with 43-53% performance improvements while ensuring all commands work as expected\!